refactor: Move strapi fetching into data-access-lib
