+++ libs/shared/data-access-strapi/src/index.ts
@@ -17,3 +17,4 @@
 export * from './lib/page-guard';
 export * from './lib/page.service';
 export * from './lib/universal.interceptor';
+export * from './lib/fetch-strapi-content';
